contrast

+++ neural/train_contrast.py
@@ -135,7 +135,7 @@ def sample_non_overlapping(data, start_fraction, end_fraction):
     idxs = starts + torch.rand(lengths.shape) * lengths
     print(idxs.shape)
     # idxs = torch.cat([torch.randint(start, start + length - n_samples, size=(2,)) for start, length in zip(starts, lengths)], dim=0)
-    return idxs
+    return idxs.long()
     
     idxs = []
     while len(idxs) < batch_size:
